+++ terraform/envs/prod/iam.tf
@@ -55,10 +55,14 @@ resource "aws_iam_role_policy" "github_deploy" {
           "lambda:GetFunction",
           "lambda:GetFunctionConfiguration",
           "lambda:PublishVersion",
-          "lambda:CreateAlias",
-          "lambda:UpdateAlias",
         ]
         Resource = aws_lambda_function.feedback.arn
+      },
+      {
+        Sid    = "LambdaAliasProductionOnly"
+        Effect = "Allow"
+        Action = ["lambda:CreateAlias", "lambda:UpdateAlias"]
+        Resource = "${aws_lambda_function.feedback.arn}:production"
       }
     ]
   })